+++ app/src/main/java/com/miaofen/xiaoying/common/data/bean/response/FocusOnResponse.kt
@@ -25,7 +25,7 @@ data class FocusOnResponse(
         var avatarUrl: String?,
         var nickName: String?,
         var selfFollowUser: Boolean?,
-        var userId: Int?,
+        var userId: Long?,
         var userFollowSelf: Boolean?
     )
 }
